--- nionet/src/main/java/com/example/protocol/Packer.java
@@ -1,14 +1,0 @@
-package com.example.protocol;
-
-
-public class Packer<T> {
-
-    public T bean;
-
-    public Packer() {}
-
-    public Packer(T bean) {
-        this.bean = bean;
-    }
-
-}
